📦 chore(Dockerfile): update Dockerfile to use kaddem-0.0.10.jar 📝 chore(pom.xml): update version to 0.0.10 📝 feat(docker-compose.yml): add docker-compose.yml file
The Dockerfile has been updated to use the kaddem-0.0.10.jar file instead of kaddem-0.0.9.jar. This ensures that the latest version of the application is used when building the Docker image.

The pom.xml file has been updated to set the version to 0.0.10. This aligns the version in the pom.xml file with the version of the application.

A new docker-compose.yml file has been added. This file defines a Docker Compose configuration for running the application and a MySQL database. The application image is tagged with the IMAGE_VERSION environment variable, and the necessary environment variables for the database connection are set.