📦 chore(Dockerfile): update jar file version to 0.0.11 🔧 chore(Jenkinsfile): remove unnecessary export and update docker-compose command 🔧 chore(pom.xml): update project version to 0.0.11
The Dockerfile has been updated to use the kaddem-0.0.11.jar file instead of kaddem-0.0.10.jar. This ensures that the latest version of the application is used in the Docker image.

In the Jenkinsfile, the export of IMAGE_VERSION has been removed as it is no longer needed. The docker-compose command has been updated to use the IMAGE_VERSION environment variable directly.

The pom.xml file has been updated to set the project version to 0.0.11, matching the version of the jar file used in the Dockerfile. This ensures consistency between the project version and the deployed artifact.