📦 chore(Dockerfile): update Dockerfile to use kaddem-0.0.13.jar 🔧 chore(Jenkinsfile): update Docker image version and push to Docker Hub 🔧 chore(pom.xml): update project version to 0.0.13
The Dockerfile has been updated to use the kaddem-0.0.13.jar instead of kaddem-0.0.12.jar. This ensures that the latest version of the application is used in the Docker image.

In the Jenkinsfile, the Docker image version is now tagged as v${BUILD_NUMBER} and pushed to Docker Hub under the repository moetezkhemissi/moetezkhemissi-5infini2-g5:v${BUILD_NUMBER}. The Docker Hub credentials are configured in Jenkins and the image is pushed using the docker login and docker push commands.

The pom.xml file has been updated to set the project version to